iskin_cli/src/iskin/tools/sandbox.py
"""Sandbox Tools - Integration with ISKIN Rust Sandbox modules"""

from openhands.sdk import Action, Observation, TextContent
from openhands.sdk.tool import ToolExecutor, register_tool


class DockerAction(Action):
    """Execute command in Docker container"""
    command: str
    image: str = "ubuntu:22.04"


class DockerObservation(Observation):
    """Docker execution result"""
    output: str = ""
    exit_code: int = 0


class DockerExecutor(ToolExecutor[DockerAction, DockerObservation]):
    """Execute in Docker container"""
    
    def __call__(self, action: DockerAction, conversation=None) -> DockerObservation:
        # TODO: Integrate with Rust container.rs
        return DockerObservation(output="", exit_code=0)


class DockerTool:
    """Tool for Docker operations"""
    
    @classmethod
    def create(cls, conv_state):
        return [DockerTool.definition()]
    
    @classmethod
    def definition(cls):
        from openhands.sdk import ToolDefinition
        return ToolDefinition(
            name="docker",
            description="Execute commands in Docker container",
            action_type=DockerAction,
            observation_type=DockerObservation,
            executor=DockerExecutor(),
        )


class HealthCheckAction(Action):
    """Check system health"""
    pass


class HealthCheckObservation(Observation):
    """Health check result"""
    healthy: bool = False
    message: str = ""


class HealthCheckExecutor(ToolExecutor[HealthCheckAction, HealthCheckObservation]):
    """Check system health"""
    
    def __call__(self, action: HealthCheckAction, conversation=None) -> HealthCheckObservation:
        # TODO: Integrate with Rust self_healing.rs
        return HealthCheckObservation(healthy=True, message="OK")
