iskin_cli/pyproject.toml
[project]
name = "iskin-cli"
version = "0.1.0"
description = "ISKIN CLI - Autonomous AI Agent on OpenHands SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openhands-sdk>=0.1.0",
    "openhands-tools>=0.1.0",
]

[project.scripts]
iskin = "iskin.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
